+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 - #1122 - Add clientlib category to touchui-widgets to load in Create Page wizard.
 - #1150 - Fix the empty datetime value displayed as "invalid date" in the touchui dialog 
 - #842 - Fix issue with Environment Indicator title being reset
+- #1143 - Remove current page from result from PagesReferenceProvider
 
 ## [3.11.0] - 2017-10-18